requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
PyYAML>=6.0
tqdm>=4.62.0

# Optional (for evaluation/visualization)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# For graph datasets (DGAD)
networkx>=2.6.0

# Python version
# Requires Python >= 3.8